refactor(servers): delegate app creation to createApp factory
